requirements.txt
pytest==7.2.2
playwright==1.31.1
pytest_playwright==0.3.2
requests==2.28.2
allure-pytest==2.13.0
allure-python-commons==2.13.0